test: Moved tests inside src
